backend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.3
django-extra-fields==3.0.2
django-filter==23.5
django-import-export==3.3.6
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
et-xmlfile==1.1.0
flake8==7.0.0
idna==3.6
isort==5.13.2
MarkupPy==1.14
mccabe==0.7.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.2.0
progress==1.6
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
tablib==3.5.0
urllib3==2.1.0
xlrd==2.0.1
xlwt==1.3.0
